Move application endpoint code logic
