+++ src/routes/applicationsContentRoute.js
@@ -8,10 +8,4 @@ router.get('/((?!favicon.ico)):orgName/applications/create', registerPartials, a
 router.get('/((?!favicon.ico)):orgName/applications/:applicationid', registerPartials, applicationsController.loadApplication);
 router.get('/((?!favicon.ico)):orgName/applications/:applicationid/edit', registerPartials, applicationsController.loadApplicationForEdit);
 
-router.post('/applications', registerPartials, applicationsController.saveApplication);
-router.put('/applications/:applicationid', registerPartials, applicationsController.updateApplication);
-router.delete('/applications/:applicationid', registerPartials, applicationsController.deleteApplication);
-router.post('/applications/:applicationid/reset-throttle-policy', registerPartials, applicationsController.resetThrottlingPolicy);
-router.post('/applications/:applicationid/api-keys/:env/generate', registerPartials, applicationsController.generateAPIKeys);
-
 module.exports = router;
