--- src/utils/partials/alert.hbs
@@ -1,23 +1,0 @@
-<head>
-  <link rel="stylesheet" href="/technical-styles/alert.css" />
-  <script src="/technical-scripts/alert.js" defer></script>
-</head>
-
-<body>
-  <section>
-    <div
-      class="modal fade custom-alert"
-      id="alertModal"
-      tabindex="-1"
-      aria-hidden="true"
-    >
-      <div class="modal-dialog">
-        <div class="modal-content">
-          <div class="modal-body {{typeClass}}">
-            <p class="modal-message">{{message}}</p>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-</body>
