--- src/utils/partials/delete-confirmation.hbs
@@ -1,43 +1,0 @@
-<head>
-  <link rel='stylesheet' href='/technical-styles/delete-confirmation.css' />
-  <script src="/technical-scripts/delete-confirmation.js" defer></script> 
-</head>
-
-<body>
-  <section>
-    <div
-      class='modal fade delete-alert'
-      id='deleteConfirmation'
-      tabindex='-1'
-      aria-hidden='true'
-    >
-      <div class='modal-dialog'>
-        <div class='modal-content'>
-          <div class='modal-header'>
-            <h5 class='modal-title'>Delete Application</h5>
-          </div>
-          <div class='modal-body'>
-            <p class='modal-message'>This Application will be removed</p>
-          </div>
-          <div class='modal-footer'>
-            <button
-              type='button'
-              class='btn'
-              data-bs-dismiss='modal'
-            >
-              Cancel
-            </button>
-            <button
-              type='button'
-              class='btn'
-              data-bs-dismiss='modal'
-              onclick="deleteApplication()"
-            >
-              Confirm
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-</body>
